evaluate/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
fire>=0.5.0
tqdm>=4.65.0
praat-parselmouth>=0.4.3
pyworld>=0.3.4
transformers>=4.38.0
torchmetrics>=1.1.0
